fix: mv api from cmdb to core-services
