+++ modules/cmdb/endpoints/endpoints.go
@@ -192,7 +192,6 @@ func (e *Endpoints) Routes() []httpserver.Endpoint {
 		{Path: "/api/clusters/{idOrName}", Method: http.MethodGet, Handler: e.GetCluster},
 		{Path: "/api/clusters", Method: http.MethodGet, Handler: e.ListCluster},
 		{Path: "/api/clusters/{clusterName}", Method: http.MethodDelete, Handler: e.DeleteCluster},
-		{Path: "/api/clusters/actions/dereference", Method: http.MethodPut, Handler: e.DereferenceCluster},
 
 		// 用户相关
 		{Path: "/api/users", Method: http.MethodGet, Handler: e.ListUser},
